core: Change type of axel_t bytes_done, start_byte, and size

+++ src/axel.h
@@ -121,7 +121,7 @@ typedef struct {
 	char filename[MAX_STRING];
 	double start_time;
 	int next_state, finish_time;
-	long long bytes_done, start_byte, size;
+	size_t bytes_done, start_byte, size;
 	int bytes_per_second;
 	struct timespec delay_time;
 	int outfd;
